@@ -13,14 +13,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see<http://www.gnu.org/licenses/>.
 
-using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Diagnostics;
 
 namespace Machina.FFXIV.Tests.Utility
 {
     public class MemoryTraceListener : TraceListener
     {
-        public List<string> Messages = new List<string>();
+        public Collection<string> Messages { get; set; } = new();
 
         public override void Write(string message)
         {
