@@ -47,7 +47,7 @@ namespace Machina.Tests
                 new byte[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 }
             );
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data, data.Length);
 
@@ -64,7 +64,7 @@ namespace Machina.Tests
                 new byte[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 }
             );
 
-            IPDecoder sut = new IPDecoder(1, 1, IPProtocol.TCP);
+            IPDecoder sut = new(1, 1, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data, data.Length);
 
@@ -86,13 +86,13 @@ namespace Machina.Tests
                 new byte[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 }
             );
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data, data.Length);
 
             Assert.AreEqual(0, sut.Fragments.Count);
             Assert.AreEqual(1, TestInfrastructure.Listener.Messages.Count);
-            Assert.IsTrue(TestInfrastructure.Listener.Messages[0].Contains("protocol version is neither"));
+            Assert.IsTrue(TestInfrastructure.Listener.Messages[0].Contains("protocol version is neither", StringComparison.Ordinal));
         }
 
         /// <summary>
@@ -112,7 +112,7 @@ namespace Machina.Tests
                 true
             );
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data, data.Length);
 
@@ -136,7 +136,7 @@ namespace Machina.Tests
                 new byte[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 }
             );
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data, data.Length);
 
@@ -161,13 +161,13 @@ namespace Machina.Tests
             // override packet length
             data[2] = 0xff;
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data, data.Length);
 
             Assert.AreEqual(0, sut.Fragments.Count);
             Assert.AreEqual(1, TestInfrastructure.Listener.Messages.Count);
-            Assert.IsTrue(TestInfrastructure.Listener.Messages[0].Contains("buffer too small"));
+            Assert.IsTrue(TestInfrastructure.Listener.Messages[0].Contains("buffer too small", StringComparison.Ordinal));
         }
 
         [TestMethod]
@@ -188,7 +188,7 @@ namespace Machina.Tests
             // extend length of buffer
             Array.Resize(ref data, data.Length + 5);
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data, data.Length);
 
@@ -221,7 +221,7 @@ namespace Machina.Tests
             Array.Resize(ref data1, data1.Length + data2.Length);
             Array.Copy(data2, 0, data1, originalSize, data2.Length);
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data1, data1.Length);
 
@@ -247,7 +247,7 @@ namespace Machina.Tests
             Array.Resize(ref data1, data1.Length + data2.Length);
             Array.Copy(data2, 0, data1, originalSize, data2.Length);
 
-            IPDecoder sut = new IPDecoder(1, 1, IPProtocol.TCP);
+            IPDecoder sut = new(1, 1, IPProtocol.TCP);
             sut.FilterAndStoreData(data1, data1.Length);
 
             uint sourceIP = ConversionUtility.IPStringToUint("1.2.3.4");
@@ -294,7 +294,7 @@ namespace Machina.Tests
             Array.Resize(ref data1, data1.Length + data2.Length);
             Array.Copy(data2, 0, data1, originalSize, data2.Length);
 
-            IPDecoder sut = new IPDecoder(1, 1, IPProtocol.TCP);
+            IPDecoder sut = new(1, 1, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data1, data1.Length);
 
@@ -315,7 +315,7 @@ namespace Machina.Tests
                 new byte[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 }
             );
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data, data.Length);
             Assert.IsTrue(sut.Fragments.Count == 1);
@@ -343,7 +343,7 @@ namespace Machina.Tests
                 true
             );
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data, data.Length);
             Assert.AreEqual(1, sut.Fragments.Count);
@@ -382,7 +382,7 @@ namespace Machina.Tests
             int originalSize = data1.Length;
             Array.Resize(ref data1, data1.Length + data2.Length);
             Array.Copy(data2, 0, data1, originalSize, data2.Length);
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data1, data1.Length);
             Assert.AreEqual(2, sut.Fragments.Count);
@@ -424,7 +424,7 @@ namespace Machina.Tests
             int originalSize = data1.Length;
             Array.Resize(ref data1, data1.Length + data2.Length);
             Array.Copy(data2, 0, data1, originalSize, data2.Length);
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data1, data1.Length);
             Assert.AreEqual(2, sut.Fragments.Count);
@@ -459,7 +459,7 @@ namespace Machina.Tests
                 new byte[] { 11, 12, 13, 14, 15, 16, 17, 18, 19, 20 }
             );
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data1, data1.Length);
             Assert.AreEqual(1, sut.Fragments.Count);
@@ -504,7 +504,7 @@ namespace Machina.Tests
                 new byte[] { 11, 12, 13, 14, 15, 16, 17, 18, 19, 20 }
             );
 
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data1, data1.Length);
             Assert.AreEqual(1, sut.Fragments.Count);
@@ -563,7 +563,7 @@ namespace Machina.Tests
                 destinationIP,
                 new byte[] { 21, 22, 23, 24, 25, 26, 27, 28, 29, 30 }
             );
-            IPDecoder sut = new IPDecoder(sourceIP, destinationIP, IPProtocol.TCP);
+            IPDecoder sut = new(sourceIP, destinationIP, IPProtocol.TCP);
 
             sut.FilterAndStoreData(data1, data1.Length);
             Assert.AreEqual(1, sut.Fragments.Count);
@@ -599,7 +599,7 @@ namespace Machina.Tests
             byte[] packet1 = ConversionUtility.HexStringToByteArray("450005DC3201200080010000C0A8010617C01E2E0800AE740001003D6162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F70717273747576776162636465666768696A6B6C6D6E6F7071727374757677");
             byte[] packet2 = ConversionUtility.HexStringToByteArray("4500001C320100B980010000C0A8010617C01E2E6162636465666768");
 
-            IPDecoder decoder = new IPDecoder(
+            IPDecoder decoder = new(
                 ConversionUtility.IPStringToUint("192.168.1.6"),
                 ConversionUtility.IPStringToUint("23.192.30.46"),
                 IPProtocol.ICMP);
@@ -631,7 +631,7 @@ namespace Machina.Tests
             int size = sizeof(IPv4Header) + (data?.Length ?? 0);
             byte[] ret = new byte[size];
 
-            IPv4Header header = new IPv4Header();
+            IPv4Header header = new();
             header.version_ihl = (byte)((ipversion << 4) + (sizeof(IPv4Header) / 4));
             header.tos_ecn = 0;
             if (!bHardware)
@@ -679,7 +679,7 @@ namespace Machina.Tests
                 size += 8 - (size % 8);
             byte[] ret = new byte[size + sizeof(IPv6Header)];
 
-            IPv6Header header = new IPv6Header();
+            IPv6Header header = new();
             ushort temp = ConversionUtility.htons((ushort)((ipversion << 4) + trafficclass));
             header.version_ltc = (byte)(temp >> 8);
             header.htc_lfl = (byte)((temp & 0xff) + (flowlabel >> 8));
