@@ -21,7 +21,7 @@ namespace Machina.Tests.Utility
     [TestClass]
     public class TestInfrastructure
     {
-        public static MemoryTraceListener Listener = new MemoryTraceListener();
+        public static MemoryTraceListener Listener { get; set; } = new();
 
         [AssemblyInitialize]
         public static void AssemblyInitialize(TestContext _1)
