@@ -18,6 +18,13 @@ using Machina.Infrastructure;
 
 namespace Machina.FFXIV.Headers
 {
+    public enum CompressionType : ushort
+    {
+        None = 0x00,
+        Zlib = 0x01,
+        Oodle = 0x02
+    }
+
     [StructLayout(LayoutKind.Explicit)]
     public struct Server_BundleHeader
     {
@@ -32,21 +39,19 @@ namespace Machina.FFXIV.Headers
         [FieldOffset(16)]
         private readonly ulong _epoch;
         [FieldOffset(24)]
-        public ushort length;
-        [FieldOffset(26)]
-        public ushort unknown1;
+        public uint length;
         [FieldOffset(28)]
-        public ushort unknown2;
+        public ushort channel;
         [FieldOffset(30)]
         public ushort message_count;
         [FieldOffset(32)]
-        public ushort encoding;
+        public byte version;
+        [FieldOffset(33)]
+        public CompressionType compression;
         [FieldOffset(34)]
-        public ushort unknown3;
+        public ushort unknown1;
         [FieldOffset(36)]
-        public ushort unknown4;
-        [FieldOffset(38)]
-        public ushort unknown5;
+        public uint uncompressed_length;
 
         //public const int MinSize = 40;
 
