@@ -18,7 +18,6 @@ using System.Collections.Concurrent;
 using System.Diagnostics;
 using System.Net;
 using System.Net.Sockets;
-using Machina.Infrastructure;
 
 namespace Machina.Sockets
 {
@@ -27,9 +26,9 @@ namespace Machina.Sockets
         private readonly int BUFFER_SIZE = (1024 * 64) + 1;
 
         private readonly object _lockObject = new object();
-        private Socket _socket;
 
-        private readonly ConcurrentQueue<Tuple<byte[], int>> _pendingBuffers = new ConcurrentQueue<Tuple<byte[], int>>();
+        private Socket _socket;
+        private ConcurrentQueue<byte[]> _pendingBuffers;
         private byte[] _currentBuffer;
 
         private bool _disposedValue;
@@ -44,8 +43,8 @@ namespace Machina.Sockets
                 // create the socket
                 _socket = CreateRawSocket(localAddress, remoteAddress);
 
-                // set buffer
                 _currentBuffer = new byte[BUFFER_SIZE];
+                _pendingBuffers = new ConcurrentQueue<byte[]>();
 
                 // start receiving data asynchronously
                 _ = _socket.BeginReceive(_currentBuffer, 0, _currentBuffer.Length, SocketFlags.None, new AsyncCallback(OnReceive), null);
@@ -76,8 +75,8 @@ namespace Machina.Sockets
 
         public CapturedData Receive()
         {
-            if (_pendingBuffers.TryDequeue(out Tuple<byte[], int> next))
-                return new CapturedData { Buffer = next.Item1, Size = next.Item2 };
+            if (_pendingBuffers.TryDequeue(out byte[] next))
+                return new CapturedData { Buffer = next, Size = next.Length };
 
             return new CapturedData { Buffer = null, Size = 0 };
         }
@@ -131,17 +130,20 @@ namespace Machina.Sockets
                     if (_socket == null)
                         return;
 
-                    byte[] buffer = _currentBuffer;
+                    byte[] buffer = null;
 
                     int received = _socket.EndReceive(ar);
                     if (received > 0)
-                        _currentBuffer = new byte[BUFFER_SIZE];
+                    {
+                        buffer = new byte[received];
+                        Array.Copy(_currentBuffer, 0, buffer, 0, received);
+                    }
 
                     _ = _socket.BeginReceive(_currentBuffer, 0, _currentBuffer.Length, SocketFlags.None, new AsyncCallback(OnReceive), null);
-
-                    if (received > 0)
+                    
+                    if (buffer != null)
                     {
-                        _pendingBuffers.Enqueue(new Tuple<byte[], int>(buffer, received));
+                        _pendingBuffers.Enqueue(buffer);
                         OnDataAvailable?.Invoke();
                     }
                 }
@@ -163,7 +165,7 @@ namespace Machina.Sockets
                 _currentBuffer = null;
             }
 
-            while (_pendingBuffers.TryDequeue(out Tuple<byte[], int> _))
+            while (_pendingBuffers.TryDequeue(out byte[] _))
             {
 
             }
